requirements.txt
streamlit
pandas
plotly
databricks-sql-connector
requests
python-dotenv